requirements.txt
asgiref
certifi
charset-normalizer
Django
djangorestframework
idna
psycopg2-binary
requests
sqlparse
tzdata
urllib3
